Updated to add constraints to a triangulation

+++ encoding/mvt/prepare.go
@@ -75,8 +75,9 @@ func PrepareGeo(geo geom.Geometry, tile *geom.Extent, pixelExtent float64) geom.
 }
 
 func preparept(g geom.Point, tile *geom.Extent, pixelExtent float64) geom.Point {
-	px := int64((g.X() - tile.MinX()) / tile.XSpan() * pixelExtent)
-	py := int64((tile.MaxY() - g.Y()) / tile.YSpan() * pixelExtent)
+
+	px := (g.X() - tile.MinX()) / tile.XSpan() * pixelExtent
+	py := (tile.MaxY() - g.Y()) / tile.YSpan() * pixelExtent
 
 	return geom.Point{float64(px), float64(py)}
 }
